requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
fpdf2>=2.7.9
python-ulid>=2.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
bcrypt>=4.2.1
